changed login view from index to /login
