+++ src/styles/login.css
@@ -4,6 +4,9 @@
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
 .login-container {
@@ -16,18 +19,14 @@
   background: rgba(102, 204, 255, 0.9);
   padding: 40px;
   border-radius: 10px;
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
   box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
     0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 2px 1px -1px rgba(0, 0, 0, 0.12);
 }
 
 .login-container .login-text-container {
   display: flex;
-  text-align: center;
   justify-content: center;
+  align-items: center;
   background-color: transparent;
 }
